requirements.txt
streamlit
pandas
requests
beautifulsoup4
matplotlib
seaborn
python-dotenv
